Initial Commit
This is the initial commit for rsyslog_postgres_tools, which open sources this codebase to the
public. This includes working support for tailing rsyslog logs out of postgres, command line
searching, an HTTP log browser and API, a cleaning command, and bootstraping the default
schema with migration change control. What still needs to be done: fixing the Dockerfile and Makefile,
more tests, fix setup.py, add Travis and coveralls configuration, etc.